added argparse; added full file hashing

+++ dupinator.py
@@ -12,15 +12,22 @@ import sys
 import stat
 import md5
 import pipes
+import argparse
 from collections import defaultdict
 
-filesBySize = defaultdict(list)
-requireEqualNames = False
+
+REQUIREEQUALNAMES = False
 MINSIZE = 100
-FIRST_SCAN_BYTES = 4096# 1024
+FIRSTSCANBYTES = 4096# 1024
 FORBIDDEN = [ 'Thumbs.db', '.DS_Store' ]
+IGNORE = []
+CONTAINING = []
+
 dirsRead = 0
 filesRead = 0
+
+filesBySize = defaultdict(list)
+
 
 def walker(arg, dirname, fnames):
     d = os.getcwd()
@@ -41,6 +48,8 @@ def walker(arg, dirname, fnames):
     sys.stdout.flush()
 
     fnames = [ f for f in fnames if f not in FORBIDDEN ]
+    fnames = [ f for f in fnames if any([f.endswith(g) for g in IGNORE]) ]
+    fnames = [ f for f in fnames if any([(g in f) for g in CONTAINING]) ]
 
     for f in fnames:
         filesRead += 1
@@ -74,18 +83,52 @@ def fmt3(num):
 
 
 def main():
-    if len(sys.argv) == 1:
-        print "not path given"
+    global REQUIREEQUALNAMES
+    global MINSIZE
+    global FIRSTSCANBYTES
+    global FORBIDDEN
+    global IGNORE
+    global CONTAINING
+
+
+    parser = argparse.ArgumentParser(description='Dupinator. Fast and efficient identification of duplicated files.')
+
+    parser.add_argument('--equal', '--equal_names'     , dest="requireEqualNames", action=store_true, help="Require files to have the same name")
+    parser.add_argument('--min'  , '--min_size'        , dest="minSize"          , type=int, default=MINSIZE, help="Minimum file size ({} bytes)".format(MINSIZE))
+    parser.add_argument('--first', '--first_scan_bytes', dest="firstScanBytes"   , type=int, default=firstScanBytes, help="Bytes to be read from file for quick hashing ({} bytes)".format(firstScanBytes))
+    parser.add_argument('--forbidden', '--forbidden_files', dest="forbidden"        , action='append', default=FORBIDDEN, help="files to be ignored ({}). will replace default".format(",".join(FORBIDDEN)))
+    parser.add_argument('--ignore'   , '--ignore_extension', dest="ignore"        , action='append', default=IGNORE, help="extensions to be ignored ({}). will replace default".format(",".join(IGNORE)))
+    parser.add_argument('--containing', '--ignore_containing',dest="containing"        , action='append', default=CONTAINING, help="ignore files containing text ({}). will replace default".format(",".join(CONTAINING)))
+
+
+    parser.add_argument('folders', nargs=argparse.REMAINDER)
+
+    args = parser.parse_args()
+    
+    if len(args) == 0:
+        print "not folder given"
+        parser.print_help()
         sys.exit(1)
+
+
+    REQUIREEQUALNAMES = args.requireEqualNames
+    MINSIZE = args.minSize
+    FIRSTSCANBYTES = args.firstScanBytes
+    FORBIDDEN = args.forbidden
+    IGNORE = args.ignore
+    CONTAINING = args.containing
+
+
 
 
     infiles = sys.argv[1:]
     outdb   = "_".join(infiles).replace("/","_").replace(" ","_")
     outdbFs = outdb + ".1.filesBySize.csv"
     outdbHa = outdb + ".2.hashes.csv"
-    outdbDu = outdb + ".3.dupes.csv"
-    outdbRm = outdb + ".4.rm.csv"
-    outdbLn = outdb + ".5.ln.csv"
+    outdbPu = outdb + ".3.potential_dupes.csv"
+    outdbDu = outdb + ".4.dupes.csv"
+    outdbRm = outdb + ".5.rm.csv"
+    outdbLn = outdb + ".6.ln.csv"
 
     print "Saving to {}".format(outdb)
     sys.stdout.flush()
@@ -95,7 +138,7 @@ def main():
     global filesBySize
 
     if os.path.exists(outdbFs):
-        if os.path.exists(outdbHa) or os.path.exists(outdbDu):
+        if os.path.exists(outdbHa) or os.path.exists(outdbPu) or os.path.exists(outdbDu):
             pass
         
         else:
@@ -137,10 +180,8 @@ def main():
 
 
 
-    dupesCount = 0
-
     if os.path.exists(outdbHa):
-        if os.path.exists(outdbDu):
+        if os.path.exists(outdbPu) or os.path.exists(outdbDu):
             pass
 
         else:
@@ -164,7 +205,7 @@ def main():
                     filesBySize[filesize][hashValue] = filegroup
     
     else:
-        print "creathing hashes database"
+        print "creating hashes database"
         sys.stdout.flush()
 
         for k in sorted(filesBySize.keys()):
@@ -181,23 +222,7 @@ def main():
 
             for fileName in inFiles:
                 with file(fileName, 'r') as aFile:
-                    hashValue = None
-                    if k <= FIRST_SCAN_BYTES:
-                        hasher = md5.new()
-
-                        while True:
-                            r = aFile.read(4096)
-
-                            if not len(r):
-                                break
-
-                            hasher.update(r)
-                            
-                        hashValue = hasher.hexdigest()
-                        
-                    else:
-                        hashValue = md5.new(aFile.read(FIRST_SCAN_BYTES)).hexdigest()
-                    
+                    hashValue = md5.new(aFile.read(FIRSTSCANBYTES)).hexdigest()
                     hashes[hashValue].append(fileName)  # ashearer
 
             filesBySize[k] = hashes
@@ -219,13 +244,13 @@ def main():
 
 
 
-    if os.path.exists(outdbDu):
-        print "reading dups db {}".format(outdbDu)
+    if os.path.exists(outdbPu):
+        print "reading potential dups db {}".format(outdbPu)
         sys.stdout.flush()
 
         filesBySize = defaultdict(lambda: defaultdict(list)) 
 
-        with open(outdbDu) as fhd:
+        with open(outdbPu) as fhd:
             for line in fhd:
                 line = line.strip()
                 
@@ -240,7 +265,7 @@ def main():
                 filesBySize[filesize][hashValue] = filegroup
 
     else:
-        print "creating dups database"
+        print "creating potential dups database"
         sys.stdout.flush()
 
         for k in filesBySize.keys():
@@ -254,7 +279,8 @@ def main():
                     continue
 
 
-                if requireEqualNames:
+                if REQUIREEQUALNAMES:
+                    #TODO: check if at least 2 share the same name not if all share it
                     if len(list(set([os.path.basename(f) for f in fileGroup]))) != 1:
                         del hashes[hashValue]
                         continue
@@ -262,13 +288,85 @@ def main():
             if len(hashes) == 0:
                 del filesBySize[k]
 
-        print "saving dupes db to {}".format(outdbDu)
+        print "saving potential dups db to {}".format(outdbDu)
         sys.stdout.flush()
 
+        with open(outdbPu, "w") as fhd:
+            for filesize, hashes in sorted(filesBySize.items()):
+                for hashValue, fileGroup in sorted(hashes.items()):
+                    fhd.write("{}\t{}\t{}\n".format(filesize, hashValue, "\t".join(sorted(fileGroup))))
+
+
+
+
+
+
+
+
+
+
+
+
+    if os.path.exists(outdbDu):
+        print "reading dups db {}".format(outdbDu)
+        sys.stdout.flush()
+
+        filesBySize = defaultdict(lambda: defaultdict(list)) 
+
+        with open(outdbDu) as fhd:
+            for line in fhd:
+                line = line.strip()
+                    
+                if len(line) == 0:
+                    continue
+
+                cols = line.split("\t")
+                filesize, hashValue = cols[:2]
+                filegroup = cols[2:]
+                filesize = int(filesize)
+                    
+                filesBySize[filesize][hashValue] = filegroup
+    
+    else:
+        print "creating dups database"
+        sys.stdout.flush()
+
+        for k in sorted(filesBySize.keys()):
+            inFiles = filesBySize[k]
+            hashes = defaultdict(list)
+
+            print 'Testing {} files of size {}...'.format(len(inFiles), k)
+            sys.stdout.flush()
+
+            for fileName in inFiles:
+                with file(fileName, 'r') as aFile:
+                    hasher = md5.new()
+
+                    for chunk in iter(lambda: aFile.read(4096), b""):
+                        hasher.update(chunk)
+                                            
+                    hashValue = hasher.hexdigest()
+                    hashes[hashValue].append(fileName)  # ashearer
+
+            filesBySize[k] = hashes
+
+
+        print "saving dups db to {}".format(outdbDu)
+        sys.stdout.flush()
         with open(outdbDu, "w") as fhd:
             for filesize, hashes in sorted(filesBySize.items()):
                 for hashValue, fileGroup in sorted(hashes.items()):
                     fhd.write("{}\t{}\t{}\n".format(filesize, hashValue, "\t".join(sorted(fileGroup))))
+
+
+
+
+
+
+
+
+
+
 
 
 
@@ -289,16 +387,18 @@ def main():
     for filesize, hashes in sorted(filesBySize.items()):
         for hashValue, fileGroup in sorted(hashes.items()):
             sourceCounter += 1
+
             # Sort on length, as usually less interesting duplicates have longer names
             fileGroup.sort( lambda x,y: cmp(len(x), len(y)) )
             
             orig = fileGroup[0]
+            copies = fileGroup[1:]
             origSize = os.path.getsize(orig)
-            numCopies = len(fileGroup)-1
+            numCopies = len(copies)
             #print 'Original {} is {} and has {} copies'.format(orig, fmt3(origSize), numCopies),
             
             bytesSavedHere = 0
-            for f in fileGroup[1:]:
+            for f in sorted(copies):
                 fileCounter += 1
                 fhd_rm.write('rm {}\n'.format(pipes.quote(f)))
                 fhd_ln.write('rm {1} && ln {0} {1}\n'.format(pipes.quote(orig), pipes.quote(f)))
@@ -314,6 +414,9 @@ def main():
 
     print "Will have saved {}; {} original file(s) duplicated in {} file(s).".format(fmt3(bytesSaved), fileCounter, sourceCounter)
     sys.stdout.flush()
+
+
+
     
 if __name__ == "__main__":
     main()
